pyproject.toml
[project]
name = "yanga-core"
version = "0.0.0"
description = "Core library for C/C++ software product line engineering."
readme = "README.md"
license = "MIT"
authors = [
  { name = "cuinixam", email = "cuinixam@gmail.com" },
]
requires-python = ">=3.10"
classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Intended Audience :: Developers",
  "Natural Language :: English",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Topic :: Software Development :: Libraries",
]

dependencies = [
  "py-app-dev>=2.1,<3",
  "typer>=0,<1",
  "mashumaro>=3.13,<4",
  "clanguru>=0,<1",
  "kspl>=1,<2",
  "pick>=2.2,<3",
]
urls."Bug Tracker" = "https://github.com/yanga-project/yanga-core/issues"
urls.Changelog = "https://github.com/yanga-project/yanga-core/blob/main/CHANGELOG.md"
urls.repository = "https://github.com/yanga-project/yanga-core"

[build-system]
requires = ["poetry-core>=2,<3"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
  "pytest==9.*",
  "coverage",
  "pytest-cov==7.*",
  "pypeline-runner>=1,<2",
  "pre-commit>=4,<5",
  "ruff>=0,<1",
  # docs
  "myst-parser>=4,<5",
  "furo",
  "sphinx-copybutton>=0,<1",
  "sphinx-design>=0,<1",
  "sphinx-new-tab-link>=0,<1",
  "sphinx-rtd-size>=0.2,<1",
  "sphinxcontrib-mermaid>=1,<2",
]

[tool.ruff]
line-length = 180
lint.select = [
  "B",   # flake8-bugbear
  "D",   # flake8-docstrings
  "C4",  # flake8-comprehensions
  "S",   # flake8-bandit
  "F",   # pyflake
  "E",   # pycodestyle
  "W",   # pycodestyle
  "UP",  # pyupgrade
  "I",   # isort
  "RUF", # ruff specific
]
lint.ignore = [
  "D100", # Missing docstring in public module
  "D101", # Missing docstring in public class
  "D102", # Missing docstring in public method
  "D103", # Missing docstring in public function
  "D104", # Missing docstring in public package
  "D105", # Missing docstring in magic method
  "D203", # 1 blank line required before class docstring
  "D212", # Multi-line docstring summary should start at the first line
  "D100", # Missing docstring in public module
  "D104", # Missing docstring in public package
  "D106", # Missing docstring in public nested class
  "D107", # Missing docstring in `__init__`
  "D401", # First line of docstring should be in imperative mood
]
lint.per-file-ignores."conftest.py" = [ "D100" ]
lint.per-file-ignores."setup.py" = [ "D100" ]
lint.per-file-ignores."tests/**/*" = [
  "D100",
  "D101",
  "D102",
  "D103",
  "D104",
  "S101",
]
lint.isort.known-first-party = [ "yanga_core", "tests" ]

[tool.pyproject-fmt]
max_supported_python = "3.14"

[tool.pytest]
minversion = "9.0"
addopts = [
  "-v",
  "-Wonce",
  "-p no:doctest",
  "--cov=yanga_core",
  "--cov-report=term",
  "--cov-report=xml"
]
pythonpath = [ "src" ]

[tool.coverage]
run.branch = true
paths.source = [
  "src",
  ".tox/**/site-packages",
]
report.exclude_lines = [
  "pragma: no cover",
  "@overload",
  "if TYPE_CHECKING",
  "raise NotImplementedError",
  'if __name__ == "__main__":',
]

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
mypy_path = "src/"
no_implicit_optional = true
show_error_codes = true
warn_unreachable = true
warn_unused_ignores = true
exclude = [
  'docs/.*',
  'setup.py',
]
disable_error_code = [ "import-untyped" ]

[[tool.mypy.overrides]]
module = "tests.*"
allow_untyped_defs = true

[tool.semantic_release]
allow_zero_version = true
version_toml = [ "pyproject.toml:project.version" ]
version_variables = [
  "src/yanga_core/__init__.py:__version__",
]
build_command = """
pip install uv
uv lock
git add uv.lock
uv build
"""

[tool.semantic_release.changelog]
exclude_commit_patterns = [
  '''chore(?:\([^)]*?\))?: .+''',
  '''ci(?:\([^)]*?\))?: .+''',
  '''refactor(?:\([^)]*?\))?: .+''',
  '''style(?:\([^)]*?\))?: .+''',
  '''test(?:\([^)]*?\))?: .+''',
  '''build\((?!deps\): .+)''',
  '''Merged? .*''',
  '''Initial [Cc]ommit.*''', # codespell:ignore
]

[tool.semantic_release.changelog.environment]
keep_trailing_newline = true

[tool.semantic_release.branches.main]
match = "main"

[tool.semantic_release.branches.noop]
match = "(?!main$)"
prerelease = true
